app lista

+++ Back/src/services/contrato/DELETE/deleteContract.service.js
@@ -1,7 +1,7 @@
 const { Contrato } = require("../../../models/index.model");
 
-const deleteContractService = async (id) => {
-  const deleted = await Contrato.destroy({ where: { id } });
+const deleteContractService = async (id, adminId) => {
+  const deleted = await Contrato.destroy({ where: { id, adminId } });
   return deleted;
 };
