+++ Back/src/services/foto/DELETE/deleteFoto.service.js
@@ -1,8 +1,8 @@
 const { Foto } = require("../../../models/index.model");
 
-const deleteFotoService = async (id) => {
-    const deleted = await Foto.destroy({ where: { id } });
-    return deleted;
+const deleteFotoService = async (id, adminId) => {
+  const deleted = await Foto.destroy({ where: { id, adminId } });
+  return deleted;
 };
 
 module.exports = deleteFotoService;
